Check if directory exists before removing files in regression test runner

+++ scripts/regression_test_runner.py
@@ -217,5 +217,6 @@ else:
     print(f"New timing geometric mean: {time_b}")
 
 # nuke cached benchmark data between runs
-shutil.rmtree('duckdb_benchmark_data')
+if os.path.isdir("duckdb_benchmark_data"):
+    shutil.rmtree('duckdb_benchmark_data')
 exit(exit_code)
